refactor: decouple user repository
