+++ src/apps/server/common/guards/strategies/jwt.strategy.ts
@@ -1,17 +1,21 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Inject, Injectable, NotFoundException } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { Request } from 'express';
 import { ExtractJwt, Strategy } from 'passport-jwt';
 
 import { AuthService } from '@apps/server/auth/auth.service';
 import { UserJwtToken } from '@apps/server/auth/types/jwtToken.type';
-import { UserRepository } from '@libs/modules/database/repositories/user.repository';
+import { UserRepository } from '@libs/modules/database/repositories/user/user.interface';
 import { EnvEnum } from '@libs/modules/env/env.enum';
 import { EnvService } from '@libs/modules/env/env.service';
 
 @Injectable()
 export class JwtStrategy extends PassportStrategy(Strategy) {
-  constructor(readonly authService: AuthService, readonly envService: EnvService, private readonly userRepository: UserRepository) {
+  constructor(
+    readonly authService: AuthService,
+    readonly envService: EnvService,
+    @Inject(UserRepository) private readonly userRepository: UserRepository,
+  ) {
     super({
       jwtFromRequest: ExtractJwt.fromExtractors([
         (request: Request): string => {
