+++ src/apps/server/users/user.module.ts
@@ -4,7 +4,8 @@ import { UserController } from '@apps/server/users/user.controller';
 import { UserService } from '@apps/server/users/user.service';
 import { FeedbackRepository } from '@libs/modules/database/repositories/feedback/feedback.interface';
 import { FeedbackRepositoryImpl } from '@libs/modules/database/repositories/feedback/feedback.repository';
-import { UserRepository } from '@libs/modules/database/repositories/user.repository';
+import { UserRepository } from '@libs/modules/database/repositories/user/user.interface';
+import { UserRepositoryImpl } from '@libs/modules/database/repositories/user/user.repository';
 import { SlackModule } from '@libs/modules/slack/slack.module';
 
 @Module({
@@ -12,7 +13,10 @@ import { SlackModule } from '@libs/modules/slack/slack.module';
   controllers: [UserController],
   providers: [
     UserService,
-    UserRepository,
+    {
+      provide: UserRepository,
+      useClass: UserRepositoryImpl,
+    },
     {
       provide: FeedbackRepository,
       useClass: FeedbackRepositoryImpl,
