+++ src/libs/modules/database/repositories/user/user.interface.ts
@@ -1,0 +1,7 @@
+export interface UserRepository {
+  findById(userId: number);
+  findFirst;
+  update;
+}
+
+export const UserRepository = Symbol('UserRepository');
